Improve compile time

+++ Sources/ImageEncoding.swift
@@ -51,13 +51,13 @@ public extension ImageEncoders {
     /// A default adaptive encoder which uses best encoder available depending
     /// on the input image and its configuration.
     struct Default: ImageEncoding {
-        public var compressionQuality: CGFloat
+        public var compressionQuality: Float
 
         /// Set to `true` to switch to HEIF when it is available on the current hardware.
         /// `false` by default.
         public var isHEIFPreferred = false
 
-        public init(compressionQuality: CGFloat = 0.8) {
+        public init(compressionQuality: Float = 0.8) {
             self.compressionQuality = compressionQuality
         }
 
@@ -75,7 +75,7 @@ public extension ImageEncoders {
             } else {
                 type = .png
             }
-            let encoder = ImageEncoders.ImageIO(type: type, compressionRatio: Float(compressionQuality))
+            let encoder = ImageEncoders.ImageIO(type: type, compressionRatio: compressionQuality)
             return encoder.encode(image)
         }
     }
